Test of simplified interactive mode
We are disregarding code refactoring with this branch and instead
focusing on code generation. I've got an idea for a simpler model that I
want to test out.

The goal is to have 2 ways for adding code, and that's it. The first way
only applies to things like string and numeric literals: you just type
things in. The second way applies to everything else: autocomplete. The
editor will automatically complete your stuff for you and reject any
characters that don't result in a valid completion.

Navigation is completed by navigating between holes. Have to think about
the methods for adding lines and stuff in between.